+++ changelog.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## v3.2.0 (2017-03-29)
+
+### Major Changes
+
+- New `given` and `begin` methods inside `SpecDefinition`.
+- New `describe` method to describe the spec.
+- Added the `onFail` method to control the default behaviour of a failed expectation
+- Messages are now encapsulated in a Consumer object, bringing flexibility to 
+  do whatever you want when the expectation fails.
+- Changed the `because` to return a `Consumer` object in order to minimize the impact of migrations.
+- Added the `otherwise` method to `Expectation`.
+- Added OSGi MANIFEST
+
 ## v3.1.0 (2017-03-02)
 
 ### Major Changes
@@ -59,22 +72,22 @@ This release contains a huge facelift in Kodo. The DSL is now more simple but ye
 
 - Fix a NPE when the target is null
 
-## v1.0 (31-12-2014)
+## v1.0 (2014-12-31)
 
 None significant.
 
-## v0.4 (18-07-2014)
+## v0.4 (2014-07-18)
 
 - Package name changed to `tools.devnull.kodo`
 - groupId changed to `tools.devnull`
 
-## v0.3 (14-07-2014)
+## v0.3 (2014-07-14)
 
 - Helpers moved to `Spec` class
 - Empty predicates for common objects (arrays, maps, collections, strings, ...)
 - Added more methods to Scenario
 
-## v0.2 (12-06-2014)
+## v0.2 (2014-06-12)
 
 - DSL refactoring
 - `Should` renamed to `Spec` and a new method `should` was added to it to
@@ -82,6 +95,6 @@ None significant.
 - Messages to specs
 - No more Junit dependency for compile
 
-## v0.1 (02-06-2014)
+## v0.1 (2014-06-02)
 
 Initial release.